% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{define_qtl_regions}
\alias{define_qtl_regions}
\title{Map GWAS Hits to LD Blocks (Post-GWAS QTL Region Definition)}
\usage{
define_qtl_regions(
  gwas_results,
  blocks,
  snp_info,
  p_threshold = 5e-08,
  trait_col = "trait",
  min_snps = 3L
)
}
\arguments{
\item{gwas_results}{Data frame: SNP, CHR, POS. Optional: P, trait.}

\item{blocks}{LD block data frame from run_Big_LD_all_chr().}

\item{snp_info}{Full SNP metadata data frame.}

\item{p_threshold}{Significance threshold. Default 5e-8. NULL = use all markers.}

\item{trait_col}{Trait column name. Default "trait".}

\item{min_snps}{Minimum block SNP count. Default 3L.}
}
\value{
Data frame: block_id, CHR, start_bp, end_bp, n_snps_block,
  n_sig_markers, lead_snp, lead_p, traits, n_traits, pleiotropic.
}
\description{
Maps significant GWAS markers onto LD blocks to define QTL
  regions. Blocks with significant markers from multiple traits are flagged
  pleiotropic. Implements the approach of Tong et al. (2024).
}
\references{
Tong et al. (2024) Theor Appl Genet 137:274.
}
