% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{ldx_blocks}
\alias{ldx_blocks}
\title{Example LD Block Table}
\format{
A \code{data.frame} with 9 rows and 9 columns:
\describe{
  \item{\code{start}}{Integer. Index of the first SNP in the block
    (1-based, over the full 230-SNP set including monomorphics).}
  \item{\code{end}}{Integer. Index of the last SNP.}
  \item{\code{start.rsID}}{Character. SNP ID at the block start.}
  \item{\code{end.rsID}}{Character. SNP ID at the block end.}
  \item{\code{start.bp}}{Integer. Base-pair position of the block start.}
  \item{\code{end.bp}}{Integer. Base-pair position of the block end.}
  \item{\code{CHR}}{Character. Chromosome label.}
  \item{\code{length_bp}}{Integer. \code{end.bp - start.bp + 1}. Physical
    span of the block in base pairs.}
  \item{\code{n_snps}}{Integer. Number of polymorphic SNPs in the block
    (those that passed the MAF filter and were assigned to a clique).
    For singletons \code{n_snps = 1}. Note that \code{end - start + 1}
    may exceed \code{n_snps} when monomorphic SNPs fall within the block
    boundaries.}
}
}
\source{
Derived analytically from the founder-haplotype simulation in
  \code{data-raw/generate_example_data.R}; matches expected
  \code{run_Big_LD_all_chr(ldx_geno, ldx_snp_info, CLQcut = 0.6)} output.
}
\usage{
ldx_blocks
}
\description{
A reference LD block table for \code{\link{ldx_geno}}, matching the nine
blocks (three per chromosome) embedded in the simulation. This object is
used in examples and tests as the expected output of
\code{\link{run_Big_LD_all_chr}} on \code{ldx_geno}.
}
\examples{
data(ldx_blocks)
ldx_blocks
summarise_blocks(ldx_blocks)

\donttest{
# Compare haplotype allele frequencies between two sample groups
data(ldx_geno, ldx_snp_info)
haps <- extract_haplotypes(ldx_geno, ldx_snp_info, ldx_blocks)
ids  <- rownames(ldx_geno)
cmp  <- compare_haplotype_populations(
  haplotypes  = haps,
  group1      = ids[1:60],
  group2      = ids[61:120],
  group1_name = "cycle1",
  group2_name = "cycle2"
)
head(cmp[, c("block_id","n1","n2","FST","divergent")])
}
}
\seealso{
\code{\link{ldx_geno}}, \code{\link{run_Big_LD_all_chr}},
  \code{\link{summarise_blocks}}
}
\keyword{datasets}
