% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geno.R
\name{print.LDxBlocks_backend}
\alias{print.LDxBlocks_backend}
\title{Print Method for LDxBlocks Backend}
\usage{
\method{print}{LDxBlocks_backend}(x, ...)
}
\arguments{
\item{x}{An \code{LDxBlocks_backend} object.}

\item{...}{Further arguments passed to or from other methods (unused).}
}
\description{
Print Method for LDxBlocks Backend
}
