% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geno.R
\name{summary.LDxBlocks_backend}
\alias{summary.LDxBlocks_backend}
\title{Summary Method for LDxBlocks Backend}
\usage{
\method{summary}{LDxBlocks_backend}(object, ...)
}
\arguments{
\item{object}{An \code{LDxBlocks_backend} object.}

\item{...}{Further arguments passed to or from other methods (unused).}
}
\description{
Summary Method for LDxBlocks Backend
}
