% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{unphase_to_dosage}
\alias{unphase_to_dosage}
\title{Collapse Phased Gametes to 0/1/2 Dosage}
\usage{
unphase_to_dosage(phased_list)
}
\arguments{
\item{phased_list}{List from read_phased_vcf() or phase_with_pedigree().}
}
\value{
Numeric matrix (individuals x SNPs, 0/1/2/NA).
}
\description{
Collapse Phased Gametes to 0/1/2 Dosage
}
\keyword{internal}
