% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{read_phased_vcf}
\alias{read_phased_vcf}
\title{Read Pre-Phased VCF}
\usage{
read_phased_vcf(vcf_file, min_maf = 0, verbose = TRUE)
}
\arguments{
\item{vcf_file}{Path to phased VCF or VCF.gz with 0|1 GT fields.}

\item{min_maf}{Minimum MAF. Default 0.0.}

\item{verbose}{Logical. Default TRUE.}
}
\value{
List: hap1, hap2 (SNPs x individuals, 0/1), dosage (0/1/2), snp_info, sample_ids.
}
\description{
Read Pre-Phased VCF
}
